refactor(frontend): decouple and decompose productivity screen
- Extracted complex data fetching, filtering, prioritizing, and sorting logic from `productivity.tsx` into a new `useProductivityData` hook, eliminating "Logic Leaks".
- Decomposed the massive 858-line `ProductivityScreen` widget by extracting `ProductivityHeader`, `ProductivityEmptyState`, and `EventCard` into dedicated stateless components.
- Reduced the main file size from 858 to 346 lines.
- Improved component reusability and strictly adhered to SOLID principles (SRP) in React Native.
- All extracted components are fully typed, use explicit prop definitions, and strictly rely on the app theme structure.
