+++ frontend/src/components/productivity/ProductivityEmptyState.tsx
@@ -1,0 +1,185 @@
+import React from 'react';
+import { View, Pressable, Platform, StyleSheet } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+import { useTranslation } from 'react-i18next';
+import { AppText } from '../AppText';
+import { theme } from '@/core/theme';
+import { DESIGN_TOKENS } from '@/core/design/tokens';
+import { Tab } from '@/hooks/useProductivityData';
+
+const T = {
+  onSurface: {
+    light: DESIGN_TOKENS.colors.text,
+    mutedLight: DESIGN_TOKENS.colors.muted,
+  },
+  primary: {
+    DEFAULT: DESIGN_TOKENS.colors.primary,
+    surfaceLight: DESIGN_TOKENS.colors.primarySoft,
+  },
+  white: '#FFFFFF',
+};
+
+const S = theme.spacing;
+const R = theme.borderRadius;
+
+interface ProductivityEmptyStateProps {
+  activeTab: Tab;
+  searchQuery: string;
+  onAddNote: () => void;
+  onAddTask: () => void;
+}
+
+export const ProductivityEmptyState = React.memo(
+  ({ activeTab, searchQuery, onAddNote, onAddTask }: ProductivityEmptyStateProps) => {
+    const { t } = useTranslation();
+
+    return (
+      <View style={styles.emptyContainer}>
+        <View style={styles.emptyIconCircle}>
+          <Ionicons
+            name={
+              activeTab === 'notes'
+                ? 'document-text'
+                : activeTab === 'tasks'
+                  ? 'checkbox'
+                  : activeTab === 'today'
+                    ? 'sunny-outline'
+                    : 'planet'
+            }
+            size={42}
+            color={T.primary.DEFAULT}
+          />
+        </View>
+        <AppText variant="h3" style={styles.emptyTitle}>
+          {searchQuery
+            ? t('productivity.no_matches') || 'No matches found'
+            : activeTab === 'notes'
+              ? t('productivity.no_notes') || 'No Notes'
+              : activeTab === 'tasks'
+                ? t('productivity.no_tasks') || 'No Tasks'
+                : activeTab === 'today'
+                  ? t('productivity.nothing_urgent_today') || 'Nothing urgent today'
+                  : t('productivity.workspace_clear') || 'Your workspace is clear'}
+        </AppText>
+        <AppText style={styles.emptySubtitle}>
+          {searchQuery
+            ? t('productivity.try_adjust_search') || 'Try adjusting your search terms.'
+            : activeTab === 'today'
+              ? t('productivity.today_empty_detail') || "Enjoy the clear schedule or get ahead on upcoming tasks."
+              : t('productivity.capture_thoughts') ||
+                'Capture your thoughts and track what needs to get done.'}
+        </AppText>
+
+        {!searchQuery && (
+          <View style={styles.emptyActions}>
+            {(activeTab === 'all' || activeTab === 'notes') && (
+              <Pressable
+                onPress={onAddNote}
+                style={({ pressed }) => [styles.emptyButton, pressed && { opacity: 0.8 }]}
+              >
+                <Ionicons name="add" size={18} color={T.white} style={{ marginEnd: 6 }} />
+                <AppText style={styles.emptyButtonText}>
+                  {t('productivity.newNote') || 'New Note'}
+                </AppText>
+              </Pressable>
+            )}
+            {(activeTab === 'all' || activeTab === 'tasks' || activeTab === 'today') && (
+              <Pressable
+                onPress={onAddTask}
+                style={({ pressed }) => [
+                  styles.emptyButton,
+                  (activeTab === 'all' || activeTab === 'today') && styles.emptyButtonSecondary,
+                  pressed && { opacity: 0.8 },
+                ]}
+              >
+                <Ionicons
+                  name="add"
+                  size={18}
+                  color={activeTab === 'all' || activeTab === 'today' ? T.primary.DEFAULT : T.white}
+                  style={{ marginEnd: 6 }}
+                />
+                <AppText
+                  style={
+                    activeTab === 'all' || activeTab === 'today'
+                      ? styles.emptyButtonTextSecondary
+                      : styles.emptyButtonText
+                  }
+                >
+                  {t('productivity.new_task') || 'New Task'}
+                </AppText>
+              </Pressable>
+            )}
+          </View>
+        )}
+      </View>
+    );
+  }
+);
+
+ProductivityEmptyState.displayName = 'ProductivityEmptyState';
+
+const styles = StyleSheet.create({
+  emptyContainer: {
+    alignItems: 'center',
+    paddingVertical: S.massive,
+  },
+  emptyIconCircle: {
+    width: 80,
+    height: 80,
+    borderRadius: 40,
+    backgroundColor: T.primary.surfaceLight,
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginBottom: S.lg,
+  },
+  emptyTitle: {
+    marginBottom: S.sm,
+    textAlign: 'center',
+    color: T.onSurface.light,
+  },
+  emptySubtitle: {
+    textAlign: 'center',
+    marginBottom: S.xl,
+    color: T.onSurface.mutedLight,
+    paddingHorizontal: S.xxxl,
+    lineHeight: 22,
+  },
+  emptyActions: {
+    flexDirection: 'row',
+    gap: S.md,
+  },
+  emptyButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: T.primary.DEFAULT,
+    borderRadius: R.xl,
+    paddingVertical: S.md,
+    paddingHorizontal: S.xl,
+    ...theme.elevation.md,
+  },
+  emptyButtonSecondary: {
+    backgroundColor: T.primary.surfaceLight,
+    ...Platform.select({
+      ios: {
+        shadowOpacity: 0,
+        elevation: 0,
+      },
+      android: {
+        elevation: 0,
+      },
+      default: {
+        elevation: 0,
+      },
+    }),
+  },
+  emptyButtonText: {
+    color: T.white,
+    fontWeight: '700',
+    fontSize: 15,
+  },
+  emptyButtonTextSecondary: {
+    color: T.primary.DEFAULT,
+    fontWeight: '700',
+    fontSize: 15,
+  },
+});
